refactor: split config into files
