+++ apps/marketing/src/components/compare/ComparisonCell.astro
@@ -1,5 +1,5 @@
 ---
-import type { NormalizedCellValue } from "@blinkdisk/constants/comparison";
+import type { NormalizedCellValue } from "@blinkdisk/utils/tools";
 import CheckIcon from "@lucide/astro/icons/check";
 import XIcon from "@lucide/astro/icons/x";
 import MinusIcon from "@lucide/astro/icons/minus";
