+++ apps/marketing/src/components/compare/ComparisonRow.astro
@@ -1,5 +1,5 @@
 ---
-import type { NormalizedCellValue } from "@blinkdisk/constants/comparison";
+import type { NormalizedCellValue } from "@blinkdisk/utils/tools";
 import ComparisonCell from "@marketing/components/compare/ComparisonCell.astro";
 import WindowsIcon from "@marketing/components/icons/Windows.astro";
 import MacosIcon from "@marketing/components/icons/Macos.astro";
