+++ apps/marketing/src/components/compare/ComparisonTable.astro
@@ -1,9 +1,9 @@
 ---
+import type { LabelConfig } from "@blinkdisk/constants/comparison";
 import type {
-  LabelConfig,
   NormalizedBackupTool,
   NormalizedCellValue,
-} from "@blinkdisk/constants/comparison";
+} from "@blinkdisk/utils/tools";
 import ComparisonRow from "@marketing/components/compare/ComparisonRow.astro";
 
 type Props = {
